requirements.txt
# requirements.txt - Specific versions for Render

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# NLP
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.36.0

# Text Processing
textstat==0.7.3
python-docx==1.1.0
PyPDF2==3.0.1
rake-nltk==1.0.4

# Machine Learning
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
loguru==0.7.2
aiofiles==23.2.1
python-dotenv==1.0.0